reorganize crc32 code
