+++ libz-rs-sys/examples/crc32_bench.rs
@@ -1,23 +1,16 @@
 //! a binary just so we can look at the optimized assembly
 
-fn main() {
+pub fn main() {
     let mut it = std::env::args();
 
     let _ = it.next().unwrap();
 
     match it.next().unwrap().as_str() {
-        "generic" => {
-            let path = it.next().unwrap();
-            let input = std::fs::read(path).unwrap();
-
-            zlib_rs::crc32_generic::<5>(&input, 0);
-        }
-
         "sse" => {
             let path = it.next().unwrap();
             let input = std::fs::read(path).unwrap();
 
-            let mut state = zlib_rs::crc32_pclmulqdq::Crc32Fold::new();
+            let mut state = zlib_rs::crc32::Crc32Fold::new();
             state.fold(&input, 0);
             println!("{:#x}", state.finish());
         }
@@ -35,7 +28,7 @@ fn main() {
             let path = it.next().unwrap();
             let input = std::fs::read(path).unwrap();
 
-            let mut state = zlib_rs::crc32_pclmulqdq::Crc32Fold::new();
+            let mut state = zlib_rs::crc32::Crc32Fold::new();
 
             for c in input.chunks(32) {
                 state.fold(c, 0);
